% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classquiz-package.R
\docType{package}
\name{classquiz-package}
\alias{classquiz}
\alias{classquiz-package}
\title{classquiz: Helps Add Quizes to Quarto Presentations}
\description{
Facilitates using google forms for classroom quizes and using observable to plot the results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/richardjtelford/classquiz}
  \item Report bugs at \url{https://github.com/richardjtelford/classquiz/issues}
}

}
\author{
\strong{Maintainer}: Richard J. Telford \email{richard.telford@uib.no} (\href{https://orcid.org/0000-0001-9826-3076}{ORCID})

}
\keyword{internal}
