% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_add_form.R
\name{cq_add_form}
\alias{cq_add_form}
\title{Add google form}
\usage{
cq_add_form(form_url, height = "600px")
}
\arguments{
\item{form_url}{URL of google form}

\item{height}{height of iframe.}
}
\value{
An R object that knitr recognizes to generate the iframe.
}
\description{
Adds a google form in an iframe to a quarto document. A wrapper for \code{\link[knitr:include_url]{knitr::include_url()}} with differnt defaults.
}
\examples{
cq_add_form(
  form_url = "https://docs.google.com/forms/d/e/1FAIpQLSclFEEt2PIQhnr-laqmtSk1CTGSyv43VaroA05G978_ammHIg/viewform?embedded=true"
)

}
