% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_add_results.R
\name{cq_add_results}
\alias{cq_add_results}
\title{Add results of quiz to quarto document}
\usage{
cq_add_results(sheet_url, id = 1)
}
\arguments{
\item{sheet_url}{URL to the google sheet containing the results}

\item{id}{id of the form (arbitrary text to distinguish multiple forms).}
}
\description{
Writes observable chunks that add the results of a quiz in google forms to a quarto document.
}
\details{
Remember to delete the results from the google sheet before class.
}
