% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_ojs_load_data.R
\name{cq_ojs_load_data}
\alias{cq_ojs_load_data}
\title{Add ojs chunk to load data}
\usage{
cq_ojs_load_data(sheet_url, id)
}
\arguments{
\item{sheet_url}{URL of the google sheet with the responses to the google form.}

\item{id}{The id of the form}
}
\description{
Add ojs chunk to load data
}
