% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_add_barplot.R
\name{cq_ojs_multiple_choice_grid_barplot}
\alias{cq_ojs_multiple_choice_grid_barplot}
\title{Add ojs barplot for multiple choice grid form}
\usage{
cq_ojs_multiple_choice_grid_barplot(id)
}
\arguments{
\item{id}{The id of the form to be plotted}
}
\description{
Adds an ojs code chunk that generates a barplot suitable for a multiple choice grid google form.
}
\details{
If this plot this code generates is not suitable for your needs, you can modify the code this function produces.
}
