% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_ojs_refresh.R
\name{cq_ojs_refresh}
\alias{cq_ojs_refresh}
\title{Add a refresh button}
\usage{
cq_ojs_refresh(id, text = "Refresh")
}
\arguments{
\item{id}{The id of the form to be plotted.}

\item{text}{The text to be displayed on the button.}
}
\description{
Adds an ojs code chunk that generates a refresh button.
}
