% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_ojs_setup.R
\name{cq_ojs_setup}
\alias{cq_ojs_setup}
\title{Add ojs setup}
\usage{
cq_ojs_setup()
}
\value{
NULL. Used for its side effect.
}
\description{
Adds an ojs setup code chunk to a quarto document. This chunk imports the arquero library (equivalent to tidyverse), sets up a function to get the CSV URL from a Google Sheets URL, and adds the URL of the google sheet with the responses to the form.
}
