% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_add_results.R
\name{cq_add_results}
\alias{cq_add_results}
\title{Add results of quiz to quarto document}
\usage{
cq_add_results(sheet_url, type)
}
\arguments{
\item{sheet_url}{URL to the google sheet containing the results}
}
\description{
Writes observable chunks that add the results of a quiz in google forms to a quarto document.
}
\details{
IMPORTANT: The chunk used to run this function must have the chunk option \verb{#| output: asis}.

Remember to delete the results from the google sheet before class.
}
