% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_ojs_process_data.R
\name{cq_ojs_process_data}
\alias{cq_ojs_process_data}
\title{Add ojs chunk to load data}
\usage{
cq_ojs_process_data(id)
}
\arguments{
\item{id}{The id of the form}
}
\description{
Add ojs chunk to load data
}
